requirements.txt
langgraph==0.2.14
langchain-google-genai==1.0.10
pymongo==4.8.0
requests==2.31.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
celery==5.3.4
redis==5.0.1